requirements.txt
flask-restful==0.3.7
sqlalchemy==1.3.1
pymysql==0.9.3
Flask-Testing==0.7.1
Flask-SQLAlchemy==2.3.2